big integer bug fixed

+++ utility/src/main/java/ShortStatisticsHandler.java
@@ -8,7 +8,7 @@ public class ShortStatisticsHandler implements StatisticsHandler {
     }
 
     @Override
-    public void handleInt(int intVal) {
+    public void handleInt(Long intVal) {
         ++intElementsNumber;
     }
